feat(electron): check if apiProcess exist before killing it

+++ app/main.js
@@ -11,9 +11,9 @@ const url = require('url')
 // be closed automatically when the JavaScript object is garbage collected.
 let mainWindow
 
-function createWindow () {
+function createWindow() {
   // Create the browser window.
-  mainWindow = new BrowserWindow({width: 800, height: 600})
+  mainWindow = new BrowserWindow({ width: 800, height: 600 })
 
   // and load the index.html of the app.
   mainWindow.loadURL(url.format({
@@ -87,9 +87,11 @@ function startApi() {
 // Kill process when electron exits
 process.on('exit', function () {
   writeLog('exit');
-  apiProcess.kill();
+  if (apiProcess !== null) {
+    apiProcess.kill();
+  }
 });
 
-function writeLog(msg){
+function writeLog(msg) {
   console.log(msg);
 }
